+++ app/asset/icon/Logo.jsx
@@ -1,0 +1,90 @@
+import { useMantineTheme } from '@mantine/core';
+import useResponsive from '~/hooks/useResponsive';
+
+export function HeaderLogo({ width = 150, ...others }) {
+  const { tablet } = useResponsive();
+  return (
+    // <svg
+    //   {...others}
+    //   width={tablet ? 120 : width}
+    //   height='60'
+    //   viewBox='0 0 150 60'
+    //   fill='none'
+    //   xmlns='http://www.w3.org/2000/svg'
+    // >
+    <svg
+      {...others}
+      // width={tablet ? 120 : width}
+      width='103'
+      height='34'
+      viewBox='0 0 103 34'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path
+        d='M20.5879 1.22852H0.919922V6.7386V17.2029C0.919922 17.2029 0.919922 17.2029 0.919922 17.5052C0.919922 25.4144 7.45978 33.3235 20.5975 33.3235C33.7256 33.3235 40.2268 25.4631 40.2268 17.5052C40.2268 9.54728 33.861 1.22852 20.5879 1.22852ZM26.5666 19.1826C22.5421 19.1924 18.5369 18.6267 14.6575 17.4955L15.0348 15.2232C18.7497 16.2959 22.5808 16.8615 26.4215 16.9103L26.5666 19.1826Z'
+        fill='#040000'
+      />
+      <path
+        d='M66.3762 26.1218C66.0086 25.7805 65.5829 25.5172 65.1089 25.3416C64.6348 25.1759 64.1317 25.0978 63.6384 25.1368H52.1162C51.1875 25.1954 50.3265 25.6147 49.688 26.2974C49.0495 26.98 48.7012 27.8773 48.7012 28.8135C48.7012 29.7497 49.0495 30.6469 49.688 31.3296C50.3265 32.0123 51.1875 32.4219 52.1162 32.4901H63.6384C64.1414 32.5194 64.6348 32.4511 65.1089 32.2853C65.5829 32.1195 66.0086 31.8465 66.3762 31.5051C66.7438 31.1638 67.034 30.7445 67.2275 30.2861C67.4307 29.8277 67.5274 29.3304 67.5274 28.8232C67.5274 28.3161 67.421 27.8188 67.2275 27.3604C67.034 26.8825 66.7438 26.4729 66.3762 26.1218ZM63.0869 29.3206C62.9418 29.4669 62.758 29.5547 62.5548 29.5742V29.5547H53.074C52.9676 29.5547 52.8611 29.5449 52.7644 29.5059C52.6677 29.4669 52.5709 29.4084 52.4935 29.3304C52.4161 29.2523 52.3581 29.1646 52.3097 29.0671C52.271 28.9695 52.242 28.8622 52.242 28.755C52.2517 28.6477 52.271 28.5502 52.3194 28.4624C52.3678 28.3746 52.4258 28.2869 52.5032 28.2186C52.5806 28.1503 52.6677 28.1016 52.7644 28.0626C52.8611 28.0333 52.9676 28.0138 53.0643 28.0235H62.4774C62.5839 28.0138 62.6806 28.0333 62.787 28.0626C62.8837 28.1016 62.9708 28.1503 63.0482 28.2186C63.1256 28.2869 63.1837 28.3746 63.232 28.4624C63.2804 28.5599 63.2997 28.6574 63.2997 28.7647C63.3094 28.9793 63.232 29.1743 63.0869 29.3206Z'
+        fill='#040000'
+      />
+      <path
+        d='M67.537 21.0897H58.656V19.0709H63.0385L62.9611 20.5338H66.9856L67.5273 12.1663H48.6914V15.1602H63.319L63.2513 16.3013H48.6914V19.0709H54.5154V21.0897H48.6914V24.0641H67.537V21.0897Z'
+        fill='#040000'
+      />
+      <path
+        d='M88.4724 15.8327V12.1755H84.3318V25.0291H75.0928V23.5956H70.9521V32.4995H88.4917V19.2167H90.8619V15.8327H88.4724ZM84.3898 29.408H75.1024V27.8963H84.3898V29.408Z'
+        fill='#040000'
+      />
+      <path
+        d='M81.4974 12.166H77.5503V14.331H74.1062V12.166H70.043V22.6791H81.4974V12.166ZM77.5503 19.7729H74.1062V17.4323H77.5503V19.7729Z'
+        fill='#040000'
+      />
+      <path
+        d='M48.6869 4.232V4.776H55.2229V4.232H48.6869ZM51.6309 3.176V4.512H52.2869V3.176H51.6309ZM49.5349 6.92V7.464H54.5029V6.92H49.5349ZM49.5349 5.376V7.096H50.1989V5.376H49.5349ZM51.9589 0.6C50.4069 0.6 49.4149 1.152 49.4149 2.048C49.4149 2.952 50.4069 3.488 51.9589 3.488C53.5029 3.488 54.5029 2.952 54.5029 2.048C54.5029 1.152 53.5029 0.6 51.9589 0.6ZM51.9589 1.144C53.0949 1.144 53.8229 1.488 53.8229 2.048C53.8229 2.616 53.0949 2.952 51.9589 2.952C50.8229 2.952 50.0949 2.616 50.0949 2.048C50.0949 1.488 50.8229 1.144 51.9589 1.144ZM60.9503 0.384V7.632H61.6143V0.384H60.9503ZM61.4543 3.264V3.824H62.7983V3.264H61.4543ZM56.3423 5.336V5.888H56.9183C58.3023 5.888 59.2623 5.832 60.3903 5.632L60.3183 5.08C59.2223 5.288 58.2783 5.336 56.9183 5.336H56.3423ZM56.3263 1.064V1.608H58.9423V3.112H56.3423V5.56H56.9983V3.656H59.5983V1.064H56.3263ZM68.6697 0.4V5.68H69.3417V0.4H68.6697ZM64.6937 6.92V7.464H69.5577V6.92H64.6937ZM64.6937 5.136V7.104H65.3497V5.136H64.6937ZM65.4617 0.896C64.3817 0.896 63.5737 1.632 63.5737 2.68C63.5737 3.72 64.3817 4.464 65.4617 4.464C66.5417 4.464 67.3417 3.72 67.3417 2.68C67.3417 1.632 66.5417 0.896 65.4617 0.896ZM65.4617 1.472C66.1657 1.472 66.7017 1.968 66.7017 2.68C66.7017 3.384 66.1657 3.88 65.4617 3.88C64.7577 3.88 64.2217 3.384 64.2217 2.68C64.2217 1.968 64.7577 1.472 65.4617 1.472ZM76.2992 1.56V2.104H78.3872V1.56H76.2992ZM76.2992 3V3.536H78.4032V3H76.2992ZM78.1312 0.392V4.672H78.7952V0.392H78.1312ZM73.1792 0.936V4.152H76.4832V0.936H73.1792ZM75.8272 1.472V3.624H73.8272V1.472H75.8272ZM76.4192 4.88C74.9232 4.88 74.0032 5.384 74.0032 6.248C74.0032 7.112 74.9232 7.608 76.4192 7.608C77.9072 7.608 78.8272 7.112 78.8272 6.248C78.8272 5.384 77.9072 4.88 76.4192 4.88ZM76.4192 5.408C77.5072 5.408 78.1712 5.72 78.1712 6.248C78.1712 6.776 77.5072 7.088 76.4192 7.088C75.3312 7.088 74.6672 6.776 74.6672 6.248C74.6672 5.72 75.3312 5.408 76.4192 5.408ZM81.0345 3.768V4.312H86.0025V3.768H81.0345ZM80.2105 6.16V6.72H86.7625V6.16H80.2105ZM83.1385 4V6.384H83.8025V4H83.1385ZM81.0345 0.968V4.008H81.6985V1.512H85.9305V0.968H81.0345ZM87.5699 6.136V6.688H94.1219V6.136H87.5699ZM90.4819 4.384V6.32H91.1379V4.384H90.4819ZM90.4659 0.872V1.432C90.4659 2.624 89.0979 3.688 87.8259 3.92L88.1059 4.472C89.5619 4.168 91.0419 2.976 91.0419 1.432V0.872H90.4659ZM90.5859 0.872V1.432C90.5859 2.976 92.0659 4.168 93.5219 4.472L93.8099 3.92C92.5379 3.688 91.1619 2.632 91.1619 1.432V0.872H90.5859ZM97.8573 2.92V4.272H98.5133V2.92H97.8573ZM97.8493 0.504V0.824C97.8493 1.792 96.6093 2.592 95.3133 2.768L95.5693 3.296C97.0413 3.056 98.4253 2.128 98.4253 0.824V0.504H97.8493ZM97.9533 0.504V0.824C97.9533 2.128 99.3453 3.048 100.809 3.296L101.057 2.768C99.7693 2.592 98.5213 1.792 98.5213 0.824V0.504H97.9533ZM94.9213 3.984V4.512H101.457V3.984H94.9213ZM98.1853 5.112C96.6333 5.112 95.7053 5.56 95.7053 6.36C95.7053 7.168 96.6333 7.608 98.1853 7.608C99.7373 7.608 100.665 7.168 100.665 6.36C100.665 5.56 99.7373 5.112 98.1853 5.112ZM98.1853 5.624C99.3293 5.624 99.9933 5.888 99.9933 6.36C99.9933 6.832 99.3293 7.096 98.1853 7.096C97.0493 7.096 96.3773 6.832 96.3773 6.36C96.3773 5.888 97.0493 5.624 98.1853 5.624Z'
+        fill='#040000'
+      />
+    </svg>
+  );
+}
+
+export function Logo({ variant = 'default', width = 30, ...others }) {
+  const theme = useMantineTheme();
+
+  return (
+    <svg
+      {...others}
+      width={width}
+      height={width * 0.8}
+      viewBox='0 0 30 24'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path
+        d='M15.0111 0H0V4.12033V11.9453C0 11.9453 0 11.9453 0 12.1714C0 18.0857 4.99139 24 15.0185 24C25.0381 24 30 18.1222 30 12.1714C30 6.2206 25.1415 0 15.0111 0ZM19.5742 13.4257C16.5026 13.433 13.4457 13.01 10.4849 12.1641L10.7728 10.4649C13.6082 11.2671 16.5321 11.6901 19.4634 11.7265L19.5742 13.4257Z'
+        fill={variant === 'default' ? theme.colors[theme.primaryColor] : '#FFF'}
+      />
+    </svg>
+  );
+}
+
+export function LogoSmall({ variant = 'default', width = 12, ...others }) {
+  const theme = useMantineTheme();
+  console.log('?', theme.colors[theme.primaryColor]);
+  return (
+    <svg
+      {...others}
+      width='12'
+      height='10'
+      viewBox='0 0 30 24'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path
+        d='M15.0111 0H0V4.12033V11.9453C0 11.9453 0 11.9453 0 12.1714C0 18.0857 4.99139 24 15.0185 24C25.0381 24 30 18.1222 30 12.1714C30 6.2206 25.1415 0 15.0111 0ZM19.5742 13.4257C16.5026 13.433 13.4457 13.01 10.4849 12.1641L10.7728 10.4649C13.6082 11.2671 16.5321 11.6901 19.4634 11.7265L19.5742 13.4257Z'
+        fill={variant === 'default' ? theme.colors[theme.primaryColor] : '#FFF'}
+      />
+    </svg>
+  );
+}
