--- app/components/common/Logo.jsx
@@ -1,81 +1,0 @@
-import { useMantineTheme } from '@mantine/core';
-import useResponsive from '~/hooks/useResponsive';
-
-export function HeaderLogo({ width = 150, ...others }) {
-  const { tablet } = useResponsive();
-  return (
-    <svg
-      {...others}
-      width={tablet ? 120 : width}
-      height='60'
-      viewBox='0 0 150 60'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M35.7919 12.4741H12V19.0862V31.6434C12 31.6434 12 31.6434 12 32.0061C12 41.4971 19.9111 50.9881 35.8036 50.9881C51.6843 50.9881 59.5486 41.5557 59.5486 32.0061C59.5486 22.4566 51.8481 12.4741 35.7919 12.4741ZM43.0242 34.019C38.1558 34.0307 33.3109 33.352 28.618 31.9944L29.0744 29.2677C33.5683 30.555 38.2026 31.2338 42.8487 31.2923L43.0242 34.019Z'
-        fill='#040000'
-      />
-      <path
-        d='M91.1812 42.3462C90.7365 41.9366 90.2216 41.6206 89.6482 41.41C89.0747 41.211 88.4662 41.1174 87.8693 41.1642H73.9313C72.8078 41.2344 71.7663 41.7377 70.9939 42.5569C70.2215 43.3761 69.8002 44.4527 69.8002 45.5762C69.8002 46.6997 70.2215 47.7763 70.9939 48.5955C71.7663 49.4147 72.8078 49.9062 73.9313 49.9881H87.8693C88.4779 50.0233 89.0747 49.9413 89.6482 49.7424C90.2216 49.5434 90.7365 49.2158 91.1812 48.8062C91.626 48.3966 91.977 47.8933 92.2111 47.3433C92.4569 46.7933 92.5739 46.1964 92.5739 45.5879C92.5739 44.9793 92.4452 44.3825 92.2111 43.8325C91.977 43.259 91.626 42.7675 91.1812 42.3462ZM87.2023 46.1847C87.0267 46.3603 86.8044 46.4656 86.5586 46.489V46.4656H75.0899C74.9611 46.4656 74.8324 46.4539 74.7154 46.4071C74.5983 46.3603 74.4813 46.2901 74.3877 46.1964C74.2941 46.1028 74.2238 45.9975 74.1653 45.8805C74.1185 45.7634 74.0834 45.6347 74.0834 45.506C74.0951 45.3772 74.1185 45.2602 74.177 45.1549C74.2355 45.0496 74.3058 44.9442 74.3994 44.8623C74.493 44.7804 74.5983 44.7219 74.7154 44.6751C74.8324 44.64 74.9611 44.6165 75.0781 44.6283H86.465C86.5937 44.6165 86.7108 44.64 86.8395 44.6751C86.9565 44.7219 87.0618 44.7804 87.1555 44.8623C87.2491 44.9442 87.3193 45.0496 87.3778 45.1549C87.4363 45.2719 87.4597 45.3889 87.4597 45.5177C87.4714 45.7751 87.3778 46.0092 87.2023 46.1847Z'
-        fill='#040000'
-      />
-      <path
-        d='M92.5854 36.3077H81.8422V33.8852H87.1436L87.05 35.6406H91.9184L92.5737 25.5996H69.7883V29.1924H87.483L87.4011 30.5616H69.7883V33.8852H76.8334V36.3077H69.7883V39.8771H92.5854V36.3077Z'
-        fill='#040000'
-      />
-      <path
-        d='M117.91 29.9994V25.6108H112.902V41.0352H101.725V39.3149H96.7167V49.9995H117.934V34.0603H120.801V29.9994H117.91ZM112.972 46.2897H101.737V44.4758H112.972V46.2897Z'
-        fill='#040000'
-      />
-      <path
-        d='M109.473 25.5996H104.698V28.1976H100.532V25.5996H95.6167V38.2153H109.473V25.5996ZM104.698 34.7278H100.532V31.9191H104.698V34.7278Z'
-        fill='#040000'
-      />
-      <path
-        d='M69.7983 16.54V17.22H77.9684V16.54H69.7983ZM73.4783 15.22V16.89H74.2983V15.22H73.4783ZM70.8583 19.9V20.58H77.0683V19.9H70.8583ZM70.8583 17.97V20.12H71.6883V17.97H70.8583ZM73.8883 12C71.9483 12 70.7083 12.69 70.7083 13.81C70.7083 14.94 71.9483 15.61 73.8883 15.61C75.8183 15.61 77.0683 14.94 77.0683 13.81C77.0683 12.69 75.8183 12 73.8883 12ZM73.8883 12.68C75.3083 12.68 76.2184 13.11 76.2184 13.81C76.2184 14.52 75.3083 14.94 73.8883 14.94C72.4683 14.94 71.5583 14.52 71.5583 13.81C71.5583 13.11 72.4683 12.68 73.8883 12.68ZM85.1276 11.73V20.79H85.9576V11.73H85.1276ZM85.7576 15.33V16.03H87.4376V15.33H85.7576ZM79.3676 17.92V18.61H80.0876C81.8176 18.61 83.0176 18.54 84.4276 18.29L84.3376 17.6C82.9676 17.86 81.7876 17.92 80.0876 17.92H79.3676ZM79.3476 12.58V13.26H82.6176V15.14H79.3676V18.2H80.1876V15.82H83.4376V12.58H79.3476ZM94.7768 11.75V18.35H95.6168V11.75H94.7768ZM89.8068 19.9V20.58H95.8868V19.9H89.8068ZM89.8068 17.67V20.13H90.6268V17.67H89.8068ZM90.7668 12.37C89.4168 12.37 88.4068 13.29 88.4068 14.6C88.4068 15.9 89.4168 16.83 90.7668 16.83C92.1168 16.83 93.1168 15.9 93.1168 14.6C93.1168 13.29 92.1168 12.37 90.7668 12.37ZM90.7668 13.09C91.6468 13.09 92.3168 13.71 92.3168 14.6C92.3168 15.48 91.6468 16.1 90.7668 16.1C89.8868 16.1 89.2168 15.48 89.2168 14.6C89.2168 13.71 89.8868 13.09 90.7668 13.09ZM104.314 13.2V13.88H106.924V13.2H104.314ZM104.314 15V15.67H106.944V15H104.314ZM106.604 11.74V17.09H107.434V11.74H106.604ZM100.414 12.42V16.44H104.544V12.42H100.414ZM103.724 13.09V15.78H101.224V13.09H103.724ZM104.464 17.35C102.594 17.35 101.444 17.98 101.444 19.06C101.444 20.14 102.594 20.76 104.464 20.76C106.324 20.76 107.474 20.14 107.474 19.06C107.474 17.98 106.324 17.35 104.464 17.35ZM104.464 18.01C105.824 18.01 106.654 18.4 106.654 19.06C106.654 19.72 105.824 20.11 104.464 20.11C103.104 20.11 102.274 19.72 102.274 19.06C102.274 18.4 103.104 18.01 104.464 18.01ZM110.233 15.96V16.64H116.443V15.96H110.233ZM109.203 18.95V19.65H117.393V18.95H109.203ZM112.863 16.25V19.23H113.693V16.25H112.863ZM110.233 12.46V16.26H111.063V13.14H116.353V12.46H110.233ZM118.402 18.92V19.61H126.592V18.92H118.402ZM122.042 16.73V19.15H122.862V16.73H122.042ZM122.022 12.34V13.04C122.022 14.53 120.312 15.86 118.722 16.15L119.072 16.84C120.892 16.46 122.742 14.97 122.742 13.04V12.34H122.022ZM122.172 12.34V13.04C122.172 14.97 124.022 16.46 125.842 16.84L126.202 16.15C124.612 15.86 122.892 14.54 122.892 13.04V12.34H122.172ZM131.261 14.9V16.59H132.081V14.9H131.261ZM131.251 11.88V12.28C131.251 13.49 129.701 14.49 128.081 14.71L128.401 15.37C130.241 15.07 131.971 13.91 131.971 12.28V11.88H131.251ZM131.381 11.88V12.28C131.381 13.91 133.121 15.06 134.951 15.37L135.261 14.71C133.651 14.49 132.091 13.49 132.091 12.28V11.88H131.381ZM127.591 16.23V16.89H135.761V16.23H127.591ZM131.671 17.64C129.731 17.64 128.571 18.2 128.571 19.2C128.571 20.21 129.731 20.76 131.671 20.76C133.611 20.76 134.771 20.21 134.771 19.2C134.771 18.2 133.611 17.64 131.671 17.64ZM131.671 18.28C133.101 18.28 133.931 18.61 133.931 19.2C133.931 19.79 133.101 20.12 131.671 20.12C130.251 20.12 129.411 19.79 129.411 19.2C129.411 18.61 130.251 18.28 131.671 18.28Z'
-        fill='black'
-      />
-    </svg>
-  );
-}
-
-export function Logo({ variant = 'default', width = 30, ...others }) {
-  const theme = useMantineTheme();
-
-  return (
-    <svg
-      {...others}
-      width={width}
-      height={width * 0.8}
-      viewBox='0 0 30 24'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M15.0111 0H0V4.12033V11.9453C0 11.9453 0 11.9453 0 12.1714C0 18.0857 4.99139 24 15.0185 24C25.0381 24 30 18.1222 30 12.1714C30 6.2206 25.1415 0 15.0111 0ZM19.5742 13.4257C16.5026 13.433 13.4457 13.01 10.4849 12.1641L10.7728 10.4649C13.6082 11.2671 16.5321 11.6901 19.4634 11.7265L19.5742 13.4257Z'
-        fill={variant === 'default' ? theme.colors[theme.primaryColor] : '#FFF'}
-      />
-    </svg>
-  );
-}
-
-export function LogoSmall({ variant = 'default', width = 12, ...others }) {
-  const theme = useMantineTheme();
-  console.log('?', theme.colors[theme.primaryColor]);
-  return (
-    <svg
-      {...others}
-      width='12'
-      height='10'
-      viewBox='0 0 30 24'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M15.0111 0H0V4.12033V11.9453C0 11.9453 0 11.9453 0 12.1714C0 18.0857 4.99139 24 15.0185 24C25.0381 24 30 18.1222 30 12.1714C30 6.2206 25.1415 0 15.0111 0ZM19.5742 13.4257C16.5026 13.433 13.4457 13.01 10.4849 12.1641L10.7728 10.4649C13.6082 11.2671 16.5321 11.6901 19.4634 11.7265L19.5742 13.4257Z'
-        fill={variant === 'default' ? theme.colors[theme.primaryColor] : '#FFF'}
-      />
-    </svg>
-  );
-}
